tmux: extract #1158 alt-buffer latch to helper (PR #1431 hygiene)
The #1158 server-truth alt-buffer fix pushed TmuxSessionViewModel.kt
411 bytes over its oversized-file hygiene baseline. Move the two added
functions (latchAltBufferAgentsFromParsed / refreshAltBufferAgentPaneIds)
verbatim into a new same-package TmuxAltBufferAgentLatch object; the VM
keeps thin private forwarders. Pure behaviour-preserving move — no logic
change. Baseline down-ratcheted 909020 -> 907910.
